add some more modularity
